+++ app/assets/stylesheets/main.css.erb
@@ -639,7 +639,7 @@ input, select, textarea {
 
 		.image:before {
 			pointer-events: none;
-			background-image: image-url("overlay.png");
+			background-image: url("<%= asset_path('overlay.png') %>");
 			background-color: rgba(19, 21, 25, 0.5);
 			border-radius: 4px;
 			content: '';
@@ -1132,7 +1132,7 @@ input, select, textarea {
 			-webkit-transition-delay: 0.75s;
 			-ms-transition-delay: 0.75s;
 			transition-delay: 0.75s;
-			background-image: linear-gradient(to top, rgba(19, 21, 25, 0.5), rgba(19, 21, 25, 0.5)), image-url("overlay.png");
+			background-image: linear-gradient(to top, rgba(19, 21, 25, 0.5), rgba(19, 21, 25, 0.5)), url("<%= asset_path('overlay.png') %>");
 			background-size: auto,
  256px 256px;
 			background-position: center,
@@ -1151,7 +1151,7 @@ input, select, textarea {
 			-webkit-transition: -webkit-transform 0.325s ease-in-out, -webkit-filter 0.325s ease-in-out;
 			-ms-transition: -ms-transform 0.325s ease-in-out, -ms-filter 0.325s ease-in-out;
 			transition: transform 0.325s ease-in-out, filter 0.325s ease-in-out;
-			background-image: image-url("neneimage1.jpeg");
+			background-image: url("<%= asset_path('nenetop.jpeg') %>");
 			background-position: center;
 			background-size: cover;
 			background-repeat: no-repeat;
