+++ config/webpack/environment.js
@@ -3,7 +3,7 @@ const { environment } = require('@rails/webpacker')
 const webpack = require('webpack')
 environment.plugins.prepend('Provide',
   new webpack.ProvidePlugin({
-    $: 'jquery/src/jquery',
+    $: './jquery.min',
     jQuery: 'jquery/src/jquery',
     Popper: ['popper.js', 'default']
   })
